feat(handlers/, httpserver/, templates/): Added Update logic, functionality and edit peer page.
However, this feature is not yet complete as a bug was found when testing. When a peer is updated,
the table will create a new row instead of overwritting the existing one which also leads to
an issue when a duplicated peer is remove, the one left on the screen cannot be edited or
deleted.